tests/submission/Infraware/Forms/contents/Forms/meter_high.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>meter_high</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
		 <p>meter: <meter max=100 min=0 id='meter_id'></meter></p>
	</form>

	<script>
		
		var meter = document.getElementById("meter_id");
		meter.value = 50;
		meter.high = 20;

		if (typeof(meter.high) == "number") {
			test(function() {
				assert_equals(meter.high, 20, "high attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("high attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
